+++ kwplot/mpl_make.py
@@ -1,10 +1,8 @@
-# -*- coding: utf-8 -*-
 """
 DEPRECATED: use kwimage versions instead
 
 Functions used to explicitly make images as ndarrays using mpl/cv2 utilities
 """
-from __future__ import absolute_import, division, print_function, unicode_literals
 import numpy as np
 from kwimage import make_heatmask, make_vector_field, make_orimask  # NOQA
 
@@ -20,8 +18,9 @@ def make_legend_img(label_to_color, dpi=96, shape=(200, 200), mode='line',
     Makes an image of a categorical legend
 
     Args:
-        label_to_color (Dict[str, kwimage.Color]):
-            mapping from string label to the color.
+        label_to_color (Dict[str, kwimage.Color] | List[Dict]):
+            Mapping from string label to the color.
+            Or a list of dictionaries that contain the keys label and color.
 
     CommandLine:
         xdoctest -m kwplot.mpl_make make_legend_img --show